move sync logic to hook
